+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Ramsete/Pose.java
@@ -1,11 +1,11 @@
 package org.firstinspires.ftc.teamcode.Ramsete;
 
 public class Pose {
-	public int X;
-	public int Y;
+	public double X;
+	public double Y;
 	public double powerInitial;
 	public double powerFinal;
-	public Pose(int X, int Y, double powerInitial, double powerFinal) {
+	public Pose(double X, double Y, double powerInitial, double powerFinal) {
 		this.X = X;
 		this.Y = Y;
 		this.powerInitial = powerInitial;
